Split code into auth.go
